created View Recipe

+++ Cupcake-Project/Models/Recipe.cs
@@ -12,6 +12,6 @@ namespace Cupcake_Project.Models
 		public int ID { get; set; }
 		public int Product_ID { get; set; }
 		public int Ingredient_ID { get; set; }
-		public int Quantity_Required { get; set; }
+		public double Quantity_Required { get; set; }
 	}
 }
